+++ st-parranderos-jdo-est/st-parranderos-jdo-est/src/main/java/uniandes/isis2304/parranderos/negocio/CapacidadNormal.java
@@ -26,7 +26,7 @@ public class CapacidadNormal implements VOCapacidadNormal
 	/**
 	 * El valor numérico de la capacidad normal
 	 */
-	private int valor;
+	private double valor;
 	
 	/**
 	 * El aforo correspondiente a la capacidad normal
@@ -52,7 +52,7 @@ public class CapacidadNormal implements VOCapacidadNormal
 	 * @param valor - El valor numérico de la capacidad normal
 	 * @param aforo - El aforo correspondiente a la capacidad normal
 	 */
-	public CapacidadNormal (long id, int valor, int aforo) 
+	public CapacidadNormal (long id, double valor, int aforo) 
 	{
 		this.id = id;
 		this.valor = valor;
@@ -78,7 +78,7 @@ public class CapacidadNormal implements VOCapacidadNormal
 	/**
 	 * @return El valor numérico de la capacidad normal
 	 */
-	public int getValor() 
+	public double getValor() 
 	{
 		return valor;
 	}
@@ -86,7 +86,7 @@ public class CapacidadNormal implements VOCapacidadNormal
 	/**
 	 * @param valor - El nuevo valor numérico de la capacidad normal
 	 */
-	public void setValor(int valor) 
+	public void setValor(double valor) 
 	{
 		this.valor = valor;
 	}
