+++ st-parranderos-jdo-est/st-parranderos-jdo-est/src/main/java/uniandes/isis2304/parranderos/negocio/VOArea.java
@@ -27,7 +27,7 @@ public interface VOArea
 	/**
 	 * @return El valor numérico del área
 	 */
-	public int getValor();
+	public double getValor();
 
 	/**
 	 * @return El aforo correspondiente al área
